feat(24): Moved the Flow data structure into the yaes-data project
