+++ build.sbt
@@ -23,6 +23,13 @@ name := "yaes"
 val scala3Version = "3.6.4"
 scalaVersion := scala3Version
 
+lazy val `yaes-data` = project
+  .settings(
+    name         := "yaes-data",
+    scalaVersion := scala3Version,
+    libraryDependencies ++= commonDependencies
+  )
+
 lazy val `yaes-core` = project
   .settings(
     name         := "yaes-core",
@@ -31,7 +38,7 @@ lazy val `yaes-core` = project
   )
 
 lazy val yaes = (project in file("."))
-  .aggregate(`yaes-core`)
+  .aggregate(`yaes-core`, `yaes-data`)
   .settings(
     scalaVersion := scala3Version
   )
@@ -41,6 +48,7 @@ lazy val dependencies =
     val scalatestVersion = "3.2.19"
     val scalatest        = "org.scalatest" %% "scalatest" % scalatestVersion
   }
+
 
 lazy val commonDependencies = Seq(
   dependencies.scalatest % Test
